Log the recursive query problem
The recursive query problem is logged by checking before responding
whether the response should have happened already. IncomingQuery now
saves the time at which the response is supposed to be sent instead of
the time at which it was received.

There is some blurring of responsibility between peer and peer behavior,
but I can't be bothered right now.